Start the service when switching graph context on Interactive
